requirements.txt
flask==3.0.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
anthropic==0.25.0
openai==1.12.0
google-generativeai==0.4.0
flask-cors==4.0.0
flask-limiter==3.5.0
redis==5.0.1
aiohttp==3.9.3
pymongo==4.6.1
cachetools==5.3.2
tiktoken==0.5.2
pillow==10.2.0
python-multipart==0.0.9